JEFFREY SEARS - OFFICE OF THE GENERAL COUNSEL AT COLUMBIA UNIVERSITY
Read FW Read here are their commencement documents the 48TH not accessible there -- here it is attached see also page.msg

Read FW Read ALL VIDEOS - UNRETURNED AS OF CURRENT BY THE ZUCKERS MAINTAINED THEM IN THE INTERNET AND CONVERTD THEM MOV .msg

Read FW Read LOCATED THE VIDEOGRAPHER FOR THE ZUCKERS… IN THE FEDERAL RECORD .msg

Read FW THANKS FOR YOUR ASSISTANCE .msg


**********************************************************************
LEGAL NOTICE
This message (including any attachments) contains confidential information which may be legally privileged. If you are not the intended recipient of this message, you are hereby notified that any disclosure, copying, or distribution of this message, or the taking of any action based on its content, is strictly prohibited. If you received this message in error, please immediately notify the sender by e-mail and delete this message from your system. Thank you.
**********************************************************************