$$$ PEEEPING TOMS AT THE GENERAL COUNSELORS AND STONEHEART ASYLUM ++ BANK OF NEW YORK MELLON
BANK OF NEW YORK MELLON - READ RECEIPTS



Read FW Read LOCATED THE VIDEOGRAPHER FOR THE ZUCKERS… IN THE FEDERAL RECORD .msg
Read FW Read ALL VIDEOS - UNRETURNED AS OF CURRENT BY THE ZUCKERS MAINTAINED THEM IN THE INTERNET AND CONVERTD THEM MOV .msg
Read legal opinion misc .msg


**********************************************************************
LEGAL NOTICE
This message (including any attachments) contains confidential information which may be legally privileged. If you are not the intended recipient of this message, you are hereby notified that any disclosure, copying, or distribution of this message, or the taking of any action based on its content, is strictly prohibited. If you received this message in error, please immediately notify the sender by e-mail and delete this message from your system. Thank you.
**********************************************************************

Your message 

   To: Tanner, Benjamin Jay
   Subject: FW: **** legal opinion [misc.]
   Sent: Wednesday, February 16, 2022 9:03:12 PM (UTC-05:00) Eastern Time (US & Canada)

 was read on Wednesday, February 16, 2022 9:04:56 PM (UTC-05:00) Eastern Time (US & Canada).
  _____  

The information contained in this e-mail, and any attachment, is confidential and is intended solely for the use of the intended recipient. Access, copying or re-use of the e-mail or any attachment, or any information contained therein, by any other person is not authorized. If you are not the intended recipient please return the e-mail to the sender and delete it from your computer. Although we attempt to sweep e-mail and attachments for viruses, we do not guarantee that either are virus-free and accept no liability for any damage sustained as a result of viruses. 

Please refer to https://disclaimer.bnymellon.com/eu.htm for certain disclosures relating to European legal entities. Please note that we may use tracking pixels to monitor your interaction with reports and data delivered via this email. We take our data protection and privacy responsibilities seriously and our privacy notice explains how we collect, use and share personal information in the course of our business activities. It can be accessed at the privacy section of www.bnymellon.com.